pyproject.toml
[tool.poetry]
name = "digitalocean"
version = "0.1.0"
description = "The official client for interacting with the DigitalOcean API"
authors = ["API Engineering <api-engineering@digitalocean.com>"]
license = "Apache"

[tool.poetry.dependencies]
python = "^3.10"
azure-core = ">=1.24.0"
azure-identity = ">=1.5.0"
isodate = ">=0.6.1"
msrest = ">=0.7.1"
typing-extensions = ">=3.7.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
responses = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
